Refactored into just one folder
