+++ wad/dir_entry_test.go
@@ -1,4 +1,4 @@
-package directory
+package wad
 
 import "testing"
