+++ wad/header.go
@@ -1,9 +1,7 @@
-package header
+package wad
 
 import (
 	"fmt"
-
-	"github.com/tomanta/gdoom/wad"
 )
 
 type Header struct {
@@ -16,17 +14,17 @@ func NewHeaderFromBytes(data []byte) (Header, error) {
 	if len(data) < 12 {
 		return Header{}, fmt.Errorf("file too small: %d", len(data))
 	}
-	wadType, err := wad.StringFromBytes(data[0:4], 4)
+	wadType, err := StringFromBytes(data[0:4], 4)
 	if err != nil {
 		return Header{}, fmt.Errorf("could not read file type: %v", err)
 	}
 
-	numLumps, err := wad.Int32FromBytes(data[4:8])
+	numLumps, err := Int32FromBytes(data[4:8])
 	if err != nil {
 		return Header{}, fmt.Errorf("could not read lump count: %v", err)
 	}
 
-	dirPos, err := wad.Int32FromBytes(data[8:12])
+	dirPos, err := Int32FromBytes(data[8:12])
 	if err != nil {
 		return Header{}, fmt.Errorf("could not read directory position: %v", err)
 	}
